requirements.txt
nltk
pymongo
scikit-learn
streamlit-option-menu
numpy